# core/services/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/gcp/iam/groups": {
            "get": {
                "description": "Retrieves a list of IAM GCP groups using the Google Cloud IAM service.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of IAM GCP groups",
                "operationId": "gcp-list-groups",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_responses_gcp.GroupDetails"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing IAM group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update Group",
                "operationId": "update-group",
                "parameters": [
                    {
                        "description": "Updated group details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/iam.UpdateGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_responses_gcp.GroupDetails"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new GCP IAM group.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create GCP Group",
                "operationId": "create-group",
                "parameters": [
                    {
                        "description": "Group details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/iam.CreateGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_responses_gcp.GroupDetails"
                        }
                    }
                }
            }
        },
        "/gcp/iam/groups/{groupName}": {
            "get": {
                "description": "Get gcp-IAM-group details by group name",
                "produces": [
                    "application/json"
                ],
                "summary": "Get gcp-Group by Name",
                "operationId": "get-gcp-group-by-name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group Name",
                        "name": "groupName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_responses_gcp.GroupDetails"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an IAM group by group name",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete Group",
                "operationId": "delete-group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group Name",
                        "name": "groupName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/gcp/iam/policy": {
            "get": {
                "description": "Get IAM policy for a project",
                "produces": [
                    "application/json"
                ],
                "summary": "Get IAM Policy",
                "operationId": "gcp-one-policy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "projectID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cloudresourcemanager.Policy"
                        }
                    }
                }
            },
            "put": {
                "description": "Set IAM policy for a project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Set IAM Policy",
                "operationId": "set-iam-policy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "projectID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "IAM policy details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_requests_gcp_iam.SetIamPolicyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "IAM policy set successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete IAM policy for a project",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete IAM Policy",
                "operationId": "delete-iam-policy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "projectID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "IAM policy deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "cloudresourcemanager.AuditConfig": {
            "type": "object",
            "properties": {
                "auditLogConfigs": {
                    "description": "AuditLogConfigs: The configuration for logging of each type of\npermission.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cloudresourcemanager.AuditLogConfig"
                    }
                },
                "service": {
                    "description": "Service: Specifies a service that will be enabled for audit logging.\nFor example, `storage.googleapis.com`, `cloudsql.googleapis.com`.\n`allServices` is a special value that covers all services.",
                    "type": "string"
                }
            }
        },
        "cloudresourcemanager.AuditLogConfig": {
            "type": "object",
            "properties": {
                "exemptedMembers": {
                    "description": "ExemptedMembers: Specifies the identities that do not cause logging\nfor this type of permission. Follows the same format of\nBinding.members.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "logType": {
                    "description": "LogType: The log type that this config enables.\n\nPossible values:\n  \"LOG_TYPE_UNSPECIFIED\" - Default case. Should never be this.\n  \"ADMIN_READ\" - Admin reads. Example: CloudIAM getIamPolicy\n  \"DATA_WRITE\" - Data writes. Example: CloudSQL Users create\n  \"DATA_READ\" - Data reads. Example: CloudSQL Users list",
                    "type": "string"
                }
            }
        },
        "cloudresourcemanager.Binding": {
            "type": "object",
            "properties": {
                "condition": {
                    "description": "Condition: The condition that is associated with this binding. If the\ncondition evaluates to `true`, then this binding applies to the\ncurrent request. If the condition evaluates to `false`, then this\nbinding does not apply to the current request. However, a different\nrole binding might grant the same role to one or more of the\nprincipals in this binding. To learn which resources support\nconditions in their IAM policies, see the IAM documentation\n(https://cloud.google.com/iam/help/conditions/resource-policies).",
                    "allOf": [
                        {
                            "$ref": "#/definitions/cloudresourcemanager.Expr"
                        }
                    ]
                },
                "members": {
                    "description": "Members: Specifies the principals requesting access for a Google\nCloud resource. `members` can have the following values: *\n`allUsers`: A special identifier that represents anyone who is on the\ninternet; with or without a Google account. *\n`allAuthenticatedUsers`: A special identifier that represents anyone\nwho is authenticated with a Google account or a service account. Does\nnot include identities that come from external identity providers\n(IdPs) through identity federation. * `user:{emailid}`: An email\naddress that represents a specific Google account. For example,\n`alice@example.com` . * `serviceAccount:{emailid}`: An email address\nthat represents a Google service account. For example,\n`my-other-app@appspot.gserviceaccount.com`. *\n`serviceAccount:{projectid}.svc.id.goog[{namespace}/{kubernetes-sa}]`:\n An identifier for a Kubernetes service account\n(https://cloud.google.com/kubernetes-engine/docs/how-to/kubernetes-service-accounts).\nFor example, `my-project.svc.id.goog[my-namespace/my-kubernetes-sa]`.\n* `group:{emailid}`: An email address that represents a Google group.\nFor example, `admins@example.com`. * `domain:{domain}`: The G Suite\ndomain (primary) that represents all the users of that domain. For\nexample, `google.com` or `example.com`. *\n`deleted:user:{emailid}?uid={uniqueid}`: An email address (plus\nunique identifier) representing a user that has been recently\ndeleted. For example, `alice@example.com?uid=123456789012345678901`.\nIf the user is recovered, this value reverts to `user:{emailid}` and\nthe recovered user retains the role in the binding. *\n`deleted:serviceAccount:{emailid}?uid={uniqueid}`: An email address\n(plus unique identifier) representing a service account that has been\nrecently deleted. For example,\n`my-other-app@appspot.gserviceaccount.com?uid=123456789012345678901`.\nIf the service account is undeleted, this value reverts to\n`serviceAccount:{emailid}` and the undeleted service account retains\nthe role in the binding. * `deleted:group:{emailid}?uid={uniqueid}`:\nAn email address (plus unique identifier) representing a Google group\nthat has been recently deleted. For example,\n`admins@example.com?uid=123456789012345678901`. If the group is\nrecovered, this value reverts to `group:{emailid}` and the recovered\ngroup retains the role in the binding.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "role": {
                    "description": "Role: Role that is assigned to the list of `members`, or principals.\nFor example, `roles/viewer`, `roles/editor`, or `roles/owner`.",
                    "type": "string"
                }
            }
        },
        "cloudresourcemanager.Expr": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "Description: Optional. Description of the expression. This is a\nlonger text which describes the expression, e.g. when hovered over it\nin a UI.",
                    "type": "string"
                },
                "expression": {
                    "description": "Expression: Textual representation of an expression in Common\nExpression Language syntax.",
                    "type": "string"
                },
                "location": {
                    "description": "Location: Optional. String indicating the location of the expression\nfor error reporting, e.g. a file name and a position in the file.",
                    "type": "string"
                },
                "title": {
                    "description": "Title: Optional. Title for the expression, i.e. a short string\ndescribing its purpose. This can be used e.g. in UIs which allow to\nenter the expression.",
                    "type": "string"
                }
            }
        },
        "cloudresourcemanager.Policy": {
            "type": "object",
            "properties": {
                "auditConfigs": {
                    "description": "AuditConfigs: Specifies cloud audit logging configuration for this\npolicy.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cloudresourcemanager.AuditConfig"
                    }
                },
                "bindings": {
                    "description": "Bindings: Associates a list of `members`, or principals, with a\n`role`. Optionally, may specify a `condition` that determines how and\nwhen the `bindings` are applied. Each of the `bindings` must contain\nat least one principal. The `bindings` in a `Policy` can refer to up\nto 1,500 principals; up to 250 of these principals can be Google\ngroups. Each occurrence of a principal counts towards these limits.\nFor example, if the `bindings` grant 50 different roles to\n`user:alice@example.com`, and not to any other principal, then you\ncan add another 1,450 principals to the `bindings` in the `Policy`.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cloudresourcemanager.Binding"
                    }
                },
                "etag": {
                    "description": "Etag: `etag` is used for optimistic concurrency control as a way to\nhelp prevent simultaneous updates of a policy from overwriting each\nother. It is strongly suggested that systems make use of the `etag`\nin the read-modify-write cycle to perform policy updates in order to\navoid race conditions: An `etag` is returned in the response to\n`getIamPolicy`, and systems are expected to put that etag in the\nrequest to `setIamPolicy` to ensure that their change will be applied\nto the same version of the policy. **Important:** If you use IAM\nConditions, you must include the `etag` field whenever you call\n`setIamPolicy`. If you omit this field, then IAM allows you to\noverwrite a version `3` policy with a version `1` policy, and all of\nthe conditions in the version `3` policy are lost.",
                    "type": "string"
                },
                "version": {
                    "description": "Version: Specifies the format of the policy. Valid values are `0`,\n`1`, and `3`. Requests that specify an invalid value are rejected.\nAny operation that affects conditional role bindings must specify\nversion `3`. This requirement applies to the following operations: *\nGetting a policy that includes a conditional role binding * Adding a\nconditional role binding to a policy * Changing a conditional role\nbinding in a policy * Removing any role binding, with or without a\ncondition, from a policy that includes conditions **Important:** If\nyou use IAM Conditions, you must include the `etag` field whenever\nyou call `setIamPolicy`. If you omit this field, then IAM allows you\nto overwrite a version `3` policy with a version `1` policy, and all\nof the conditions in the version `3` policy are lost. If a policy\ndoes not include any conditions, operations on that policy may\nspecify any valid version or leave the field unset. To learn which\nresources support conditions in their IAM policies, see the IAM\ndocumentation\n(https://cloud.google.com/iam/help/conditions/resource-policies).",
                    "type": "integer"
                }
            }
        },
...
        "gitea_pcp-inariam_inariam_core_services_api_requests_gcp_iam.SetIamPolicyRequest": {
            "type": "object"
        },
...
        "gitea_pcp-inariam_inariam_core_services_api_responses_gcp.GroupDetails": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "etag": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "email": {
                                "type": "string"
                            }
                        }
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "iam.CreateGroupRequest": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "iam.UpdateGroupRequest": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}